Move PaymentsGateway to src folder
