+++ MerchantPlatform.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaymentsGateway", "PaymentsGateway\PaymentsGateway.csproj", "{5102E0D3-309E-4B07-8E18-9D893C7D09D6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PaymentsGateway", "src\PaymentsGateway\PaymentsGateway.csproj", "{51057F4B-B7BB-4055-AB66-8811895EE29F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5102E0D3-309E-4B07-8E18-9D893C7D09D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5102E0D3-309E-4B07-8E18-9D893C7D09D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5102E0D3-309E-4B07-8E18-9D893C7D09D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5102E0D3-309E-4B07-8E18-9D893C7D09D6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{51057F4B-B7BB-4055-AB66-8811895EE29F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{51057F4B-B7BB-4055-AB66-8811895EE29F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{51057F4B-B7BB-4055-AB66-8811895EE29F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{51057F4B-B7BB-4055-AB66-8811895EE29F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
